fix crud

+++ library/Core/Controller/Action/Crud.php
@@ -187,7 +187,10 @@ abstract class Core_Controller_Action_Crud extends Core_Controller_Action
     public function deleteAction()
     {
         $model = $this->_loadModel();
-        $this->_helper->json( $model->delete() );
+        if ($model) {
+        	$this->_helper->json( $model->delete() );
+        }
+
     }
 
     /**
